web servertje werkt

+++ SocialDistancingAI/logger.py
@@ -1,13 +1,15 @@
 import json
 import os
+from datetime import datetime
 
 class Logger:
     def __init__(self, filename):
-        self.filename = filename
+        self.filename = filename + datetime.now().strftime("%d-%m-%Y") + ".json"
 
-    def write_log_entry(self, time = 0, violations = 0, people = 0):
+    def write_log_entry(self, date = 0, time = 0, violations = 0, people = 0):
         data = {
-                "timeStamp": time,
+                "date": date,
+                "time": time,
                 "violations": violations,
                 "people": people
             } 
@@ -15,15 +17,18 @@ class Logger:
             
         print(f"trying to put {data} in {self.filename}")
 
-        if os.path.getsize(self.filename) == 0:
+        if not os.path.exists(self.filename):
+            with open(self.filename, 'a+') as f:
+                json.dump({"entries": []}, f)
+
+        if os.path.exists(self.filename) and os.path.getsize(self.filename) == 0:
             with open(self.filename, 'w') as f:
                 json.dump({"entries": []}, f)
 
-
         with open(self.filename, 'r') as f:
-            contents = json.load(f) 
+            contents = json.load(f)
 
-        with open(self.filename, 'w') as f: 
+        with open(self.filename, 'w') as f:
             contents['entries'].append(data) 
             json.dump(contents, f)
